requirements.txt
matplotlib==3.6.2
numpy==1.23.5
PyYAML==6.0.1
torch==1.12.1
tqdm==4.65.0
